Update typescript definitions, remove old definitions fix tests
