+++ package.json
@@ -8,15 +8,14 @@
     "url": "https://github.com/YouCanBookMe/react-datetime"
   },
   "main": "./DateTime.js",
+  "typings": "./DateTime.d.ts",
   "files": [
     "DateTime.js",
-    "react-datetime.d.ts",
-    "typings/index.d.ts",
+    "DateTime.d.ts",
     "src",
     "css",
     "dist"
   ],
-  "types": "./typings/index.d.ts",
   "scripts": {
     "build:mac": "./node_modules/.bin/gulp",
     "build:win": "./node_modules/.bin/gulp.cmd",
@@ -43,7 +42,6 @@
     "react-dom": ">=0.13"
   },
   "devDependencies": {
-    "@types/react": "^0.14.49",
     "babel-core": "^6.22.1",
     "babel-jest": "^18.0.0",
     "babel-loader": "^6.2.1",
@@ -63,10 +61,9 @@
     "jest-cli": "^18.1.0",
     "jsdom": "^7.0.2",
     "mocha": "^3.2.0",
-    "moment": ">=2.16.0",
     "pre-commit": "^1.1.3",
-    "react": ">=0.13",
-    "react-addons-test-utils": ">=0.13",
+    "react": ">=15 <16",
+    "react-addons-test-utils": ">=15",
     "react-dom": ">=0.13",
     "react-test-renderer": "^15.4.2",
     "through2": "^2.0.3",
@@ -76,6 +73,8 @@
     "webpack-stream": "^3.2.0"
   },
   "dependencies": {
+    "@types/react": ">=15 <16",
+    "moment": "2.16.x",
     "object-assign": "^3.0.0",
     "react-onclickoutside": "^5.9.0"
   },
